backend/requirements.txt
fastapi
uvicorn
pillow
opencv-python-headless
pandas
requests
ultralytics